ChatBot_AI/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
gradio
pydantic
